Separate starting the cluster agent into its own file/function
